Make ViewActionRecorder generic and used throughout 3 tests.
